requirements.txt
langchain
langchain-google-genai
langchain-community
sqlalchemy
psycopg2-binary
pandas
plotly
streamlit
python-dotenv
